refactor cases with fixtures
