--- src/lexer/source-token/testing/fixtures/index.ts
@@ -1,1 +1,0 @@
-export const fakePos = { row: 0, col: 0 };
